Add Api::Response and Do notation to Api::Request
This is a full refactoring of the Request class, converting the #call
method with monad Do notation. I think this provides a more explicit
way to integrate the "expected" Exceptions, leaving a single rescue
block, where we had 3.
Also replaces the #initialize method with Dry::Initializer params.
Add a new integration example to client_spec.rb that verifies a Failure
from a Api::Response::Status payload, with a false :success entry.